+++ types/account.go
@@ -1,17 +1,17 @@
 package types
 
 import (
+	"bytes"
 	"encoding/json"
 	"fmt"
 
 	"gopkg.in/yaml.v2"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 	"github.com/cosmos/cosmos-sdk/x/auth"
 	"github.com/cosmos/cosmos-sdk/x/auth/exported"
 	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
-
-	tmamino "github.com/tendermint/tendermint/crypto/encoding/amino"
 
 	ethcmn "github.com/ethereum/go-ethereum/common"
 	ethcrypto "github.com/ethereum/go-ethereum/crypto"
@@ -53,20 +53,20 @@ func (acc EthAccount) EthAddress() ethcmn.Address {
 
 // Balance returns the balance of an account.
 func (acc EthAccount) Balance() sdk.Int {
-	return acc.GetCoins().AmountOf(DenomDefault)
+	return acc.GetCoins().AmountOf(AttoPhoton)
 }
 
 // SetBalance sets an account's balance of aphotons
 func (acc *EthAccount) SetBalance(amt sdk.Int) {
 	coins := acc.GetCoins()
-	diff := amt.Sub(coins.AmountOf(DenomDefault))
+	diff := amt.Sub(coins.AmountOf(AttoPhoton))
 	switch {
 	case diff.IsPositive():
 		// Increase coins to amount
-		coins = coins.Add(sdk.NewCoin(DenomDefault, diff))
+		coins = coins.Add(NewPhotonCoin(diff))
 	case diff.IsNegative():
 		// Decrease coins to amount
-		coins = coins.Sub(sdk.NewCoins(sdk.NewCoin(DenomDefault, diff.Neg())))
+		coins = coins.Sub(sdk.NewCoins(NewPhotonCoin(diff.Neg())))
 	default:
 		return
 	}
@@ -78,8 +78,9 @@ func (acc *EthAccount) SetBalance(amt sdk.Int) {
 
 type ethermintAccountPretty struct {
 	Address       sdk.AccAddress `json:"address" yaml:"address"`
+	EthAddress    string         `json:"eth_address" yaml:"eth_address"`
 	Coins         sdk.Coins      `json:"coins" yaml:"coins"`
-	PubKey        []byte         `json:"public_key" yaml:"public_key"`
+	PubKey        string         `json:"public_key" yaml:"public_key"`
 	AccountNumber uint64         `json:"account_number" yaml:"account_number"`
 	Sequence      uint64         `json:"sequence" yaml:"sequence"`
 	CodeHash      string         `json:"code_hash" yaml:"code_hash"`
@@ -89,14 +90,20 @@ type ethermintAccountPretty struct {
 func (acc EthAccount) MarshalYAML() (interface{}, error) {
 	alias := ethermintAccountPretty{
 		Address:       acc.Address,
+		EthAddress:    acc.EthAddress().String(),
 		Coins:         acc.Coins,
 		AccountNumber: acc.AccountNumber,
 		Sequence:      acc.Sequence,
 		CodeHash:      ethcmn.Bytes2Hex(acc.CodeHash),
 	}
 
+	var err error
+
 	if acc.PubKey != nil {
-		alias.PubKey = acc.PubKey.Bytes()
+		alias.PubKey, err = sdk.Bech32ifyPubKey(sdk.Bech32PubKeyTypeAccPub, acc.PubKey)
+		if err != nil {
+			return nil, err
+		}
 	}
 
 	bz, err := yaml.Marshal(alias)
@@ -109,16 +116,28 @@ func (acc EthAccount) MarshalYAML() (interface{}, error) {
 
 // MarshalJSON returns the JSON representation of an EthAccount.
 func (acc EthAccount) MarshalJSON() ([]byte, error) {
+	var ethAddress = ""
+
+	if acc.BaseAccount != nil && acc.Address != nil {
+		ethAddress = acc.EthAddress().String()
+	}
+
 	alias := ethermintAccountPretty{
 		Address:       acc.Address,
+		EthAddress:    ethAddress,
 		Coins:         acc.Coins,
 		AccountNumber: acc.AccountNumber,
 		Sequence:      acc.Sequence,
 		CodeHash:      ethcmn.Bytes2Hex(acc.CodeHash),
 	}
 
+	var err error
+
 	if acc.PubKey != nil {
-		alias.PubKey = acc.PubKey.Bytes()
+		alias.PubKey, err = sdk.Bech32ifyPubKey(sdk.Bech32PubKeyTypeAccPub, acc.PubKey)
+		if err != nil {
+			return nil, err
+		}
 	}
 
 	return json.Marshal(alias)
@@ -126,26 +145,65 @@ func (acc EthAccount) MarshalJSON() ([]byte, error) {
 
 // UnmarshalJSON unmarshals raw JSON bytes into an EthAccount.
 func (acc *EthAccount) UnmarshalJSON(bz []byte) error {
-	acc.BaseAccount = &authtypes.BaseAccount{}
-	var alias ethermintAccountPretty
+	var (
+		alias ethermintAccountPretty
+		err   error
+	)
+
 	if err := json.Unmarshal(bz, &alias); err != nil {
 		return err
 	}
 
-	if alias.PubKey != nil {
-		pubKey, err := tmamino.PubKeyFromBytes(alias.PubKey)
+	switch {
+	case !alias.Address.Empty() && alias.EthAddress != "":
+		// Both addresses provided. Verify correctness
+		ethAddress := ethcmn.HexToAddress(alias.EthAddress)
+		ethAddressFromAccAddress := ethcmn.BytesToAddress(alias.Address.Bytes())
+
+		if !bytes.Equal(ethAddress.Bytes(), alias.Address.Bytes()) {
+			err = sdkerrors.Wrapf(
+				sdkerrors.ErrInvalidAddress,
+				"expected %s, got %s",
+				ethAddressFromAccAddress.String(), ethAddress.String(),
+			)
+		}
+
+	case !alias.Address.Empty() && alias.EthAddress == "":
+		// unmarshal sdk.AccAddress only. Do nothing here
+	case alias.Address.Empty() && alias.EthAddress != "":
+		// retrieve sdk.AccAddress from ethereum address
+		ethAddress := ethcmn.HexToAddress(alias.EthAddress)
+		alias.Address = sdk.AccAddress(ethAddress.Bytes())
+	case alias.Address.Empty() && alias.EthAddress == "":
+		err = sdkerrors.Wrapf(
+			sdkerrors.ErrInvalidAddress,
+			"account must contain address in Ethereum Hex or Cosmos Bech32 format",
+		)
+	}
+
+	if err != nil {
+		return err
+	}
+
+	acc.BaseAccount = &authtypes.BaseAccount{
+		Coins:         alias.Coins,
+		Address:       alias.Address,
+		AccountNumber: alias.AccountNumber,
+		Sequence:      alias.Sequence,
+	}
+	acc.CodeHash = ethcmn.Hex2Bytes(alias.CodeHash)
+
+	if alias.PubKey != "" {
+		acc.BaseAccount.PubKey, err = sdk.GetPubKeyFromBech32(sdk.Bech32PubKeyTypeAccPub, alias.PubKey)
 		if err != nil {
 			return err
 		}
-
-		acc.BaseAccount.PubKey = pubKey
 	}
-
-	acc.BaseAccount.Coins = alias.Coins
-	acc.BaseAccount.Address = alias.Address
-	acc.BaseAccount.AccountNumber = alias.AccountNumber
-	acc.BaseAccount.Sequence = alias.Sequence
-	acc.CodeHash = ethcmn.Hex2Bytes(alias.CodeHash)
-
 	return nil
+}
+
+// String implements the fmt.Stringer interface
+func (acc EthAccount) String() string {
+	out, _ := yaml.Marshal(acc)
+	return string(out)
 }